Project: Inline kitmodel.{h,cpp} into kitoptionspage.cpp
The only user.
