+++ src/plugins/projectexplorer/projectexplorer.qbs
@@ -86,7 +86,6 @@ Project {
                 "kitfeatureprovider.h",
                 "kitmanager.cpp", "kitmanager.h",
                 "kitmanagerconfigwidget.cpp", "kitmanagerconfigwidget.h",
-                "kitmodel.cpp", "kitmodel.h",
                 "kitoptionspage.cpp", "kitoptionspage.h",
                 "ldparser.cpp", "ldparser.h",
                 "lldparser.cpp", "lldparser.h",
